It compiles!

+++ Ex3/projFiles/UserCommon/OurSattelliteView.cpp
@@ -2,7 +2,7 @@
 
 using namespace UserCommon_0000;
 
-OurSattelliteView::OurSattelliteView(const vector<vector<array<shared_ptr<matrixObject>, 3>>> gBoard, int numOfCols, int numOfRows, size_t tankX, size_t tankY):
+OurSattelliteView::OurSattelliteView(const vector<vector<array<shared_ptr<matrixObject>, 3>>> gBoard, size_t numOfCols, size_t numOfRows, size_t tankX, size_t tankY):
  gameBoard(gBoard), numOfRows(numOfRows), numOfCols(numOfCols), tankX(tankX), tankY(tankY){
     // Initialize the game board with the provided shared pointer
     // The game board is a 2D vector of arrays, where each array contains three shared pointers to matrixObject.
